# taga-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Taga API server",
        "title": "Taga API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/audit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Serves audit records to authorized administrators with filtering and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Audit"
                ],
                "summary": "Get audit records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Member TAGA ID",
                        "name": "taga_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Year (YYYY)",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Month (MM)",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Action filter",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Module filter",
                        "name": "module",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page (default 50, max 200)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/audit/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the list of user IDs (taga-ids) that have audit files for the given year/month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Audit"
                ],
                "summary": "Get audit user list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Year (YYYY)",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Month (MM)",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/init-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Initialize member password(s)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Member ID or keyword: all | none (default)",
                        "name": "memberId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/admin/login": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Admin Login",
                "parameters": [
                    {
                        "description": "Admin Login Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AdminLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AdminLoginResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AdminLoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "admin123"
                },
                "username": {
                    "type": "string",
                    "example": "admin@taga.com"
                }
            }
        },
        "handler.AdminLoginResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}